fastapi_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
ollama==0.1.5
cx_Oracle==8.3.0
python-dotenv==1.0.0
pydantic==2.5.0
prometheus-client==0.19.0

